refactor: create `AboutMe` section component
